# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "https-tunnel control plane",
    "version": "1.0.0",
    "summary": "Issues tunnel sessions and reports on them.",
    "description": "The control plane runs on the base domain, `tunnel.example.com`. It is the\nonly host that serves this API and the web UI.\n\nEvery other host under the base domain, `<label>.tunnel.example.com`, is a\ntunnel. Those hostnames serve whatever the connected client is running locally\nand have no API of their own, so nothing in this document applies to them. A\nrequest to a tunnel hostname whose client is offline is answered by the proxy\nitself with `502` and the offline page.\n\nA client uses two endpoints. `POST /api/v1/connect` registers or resumes a\nsession and returns the public URL, then `GET /api/v1/tunnel` upgrades the\nconnection to the tunnel protocol and keeps it open. The remaining endpoints\nare for inspection and cleanup.\n",
    "license": {
      "name": "MIT"
    },
    "contact": {
      "name": "https-tunnel",
      "url": "/docs"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This server, whichever host is serving this document"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "tags": [
          "sessions"
        ],
        "summary": "List every session",
        "description": "Sessions are returned newest first, connected and disconnected alike.",
        "responses": {
          "200": {
            "description": "The current session table.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "sessions"
                  ],
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "pruneSessions",
        "tags": [
          "sessions"
        ],
        "summary": "Delete every idle session older than a window",
        "description": "Removes disconnected sessions whose last activity is older than idle_for, freeing their subdomains. Connected tunnels are never removed whatever their age: a tunnel that has been quiet for a week still has a client attached, and cutting it is what deleting a single session by id is for.",
        "parameters": [
          {
            "name": "idle_for",
            "in": "query",
            "required": true,
            "description": "How long a session must have been idle to be removed. The list is fixed so that a mistyped value cannot delete more than intended.",
            "schema": {
              "type": "string",
              "enum": [
                "5m",
                "30m",
                "1h",
                "6h",
                "24h",
                "2d",
                "1w",
                "1m"
              ]
            },
            "example": "24h"
          }
        ],
        "responses": {
          "200": {
            "description": "What was removed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "idle_for",
                    "deleted",
                    "sessions"
                  ],
                  "properties": {
                    "idle_for": {
                      "type": "string"
                    },
                    "deleted": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      },
                      "description": "The sessions that were deleted, oldest first."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "idle_for was missing or not one of the accepted windows.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer <api key>`."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative to the bearer header; the same key either way."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tunnel_admin",
        "description": "Session cookie set by /api/v1/auth/login. The web UI uses this; API clients use a key."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or unrecognized API key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "invalid api key"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Session": {
        "type": "object",
        "required": [
          "session",
          "subdomain",
          "url",
          "key_name",
          "connected",
          "created_at",
          "last_seen",
          "requests"
        ],
        "properties": {
          "session": {
            "type": "string"
          },
          "subdomain": {
            "type": "string",
            "description": "Label the session owns under the base domain."
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Public URL, derived from the subdomain and the scheme this request arrived over rather than stored."
          },
          "key_name": {
            "type": "string",
            "description": "Name of the API key that owns the session."
          },
          "connected": {
            "type": "boolean",
            "description": "Whether a client is attached right now."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "description": "Last attach, detach or proxied request."
          },
          "requests": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "description": "Requests proxied since the session was created."
          },
          "remote_addr": {
            "type": "string",
            "description": "Address the client is connected from. Absent while offline."
          }
        },
        "example": {
          "session": "6wjulksypp3hbptpm5hqbn2davqbroyn",
          "subdomain": "qr8xfcb946mw",
          "url": "https://qr8xfcb946mw.tunnel.example.com",
          "key_name": "chris-laptop",
          "connected": true,
          "created_at": "2026-08-23T18:52:24.361Z",
          "last_seen": "2026-08-23T18:52:26.594Z",
          "requests": 2,
          "remote_addr": "203.0.113.17"
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}